backend/requirements.txt
# API and Web Framework
fastapi>=0.100.0
uvicorn>=0.22.0
markdown>=3.5.0

# LLM and AI
groq>=0.4.0
google-cloud-firestore>=2.11.0
google-api-python-client>=2.95.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0

# Integration SDKs
slack-sdk>=3.21.0

# Data Processing and Parsing
pandas>=2.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pymupdf>=1.22.0  # Used as 'fitz'

# Document Generation
python-docx>=0.8.11
weasyprint>=60.0  # For PDF generation from HTML/Markdown

# Database
psycopg2-binary>=2.9.0

# Testing
pytest>=7.4.0
